sportshield-ai/backend/requirements.txt
# SportShield AI | Google Solution Challenge 2026 | First Prize Target
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
imagehash==4.3.1
Pillow==10.3.0
piexif==1.1.3
numpy==1.26.4
opencv-python-headless==4.9.0.80
stegano==0.11.3
transformers==4.41.2
torch==2.3.0
torchvision==0.18.0
google-generativeai==0.7.2
firebase-admin==6.5.0
reportlab==4.2.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
beautifulsoup4==4.12.3
sendgrid==6.11.0
websockets==12.0
librosa==0.10.1
soundfile==0.12.1
APScheduler==3.10.4
slowapi==0.1.9
